backend/requirements.txt
fastapi
uvicorn
python-multipart
yt-dlp
librosa
ffmpeg-python
openai-whisper
mlflow
pandas
numpy
